+++ src/js/components/filter.js
@@ -21,7 +21,7 @@ import {
     toggleClass,
     trigger,
 } from 'uikit-util';
-import { keyMap } from '../mixin/utils';
+import { keyMap } from '../util/keys';
 
 export default {
     mixins: [Animate],