+++ src/js/core/leader.js
@@ -1,10 +1,10 @@
 import Class from '../mixin/class';
 import Media from '../mixin/media';
+import { resize } from '../api/observables';
 import { attr, css, toggleClass, unwrap, wrapInner } from 'uikit-util';
-import Resize from '../mixin/resize';
 
 export default {
-    mixins: [Class, Media, Resize],
+    mixins: [Class, Media],
 
     props: {
         fill: String,
@@ -30,6 +30,8 @@ export default {
     disconnected() {
         unwrap(this.wrapper.childNodes);
     },
+
+    observe: resize(),
 
     update: {
         read() {