+++ src/js/util/keys.js
@@ -1,0 +1,11 @@
+export const keyMap = {
+    TAB: 9,
+    ESC: 27,
+    SPACE: 32,
+    END: 35,
+    HOME: 36,
+    LEFT: 37,
+    UP: 38,
+    RIGHT: 39,
+    DOWN: 40,
+};